agent/requirements.txt
langchain-ollama
langchain-core
langchain-community
gspread
fastapi
uvicorn
pydantic
Pillow
pytesseract
fitz
python-docx
pandas
python-pptx
pymupdf
nest-asyncio
mcp
langchain-mcp-adapters